# DEV: Speed up "having working core features" shared example by ~30%