--- a/spec/system/page_objects/components/composer.rb
+++ b/spec/system/page_objects/components/composer.rb
@@ -54,10 +54,7 @@ module PageObjects
       end
 
       def fill_content(content)
-        page.driver.with_playwright_page do |pw_page|
-          pw_page.locator("#{COMPOSER_ID} .d-editor .d-editor-input").press_sequentially(content)
-        end
-
+        find("#{COMPOSER_ID} .d-editor .d-editor-input").fill_in(with: content)
         self
       end